[feat,Refactor] 리펙토링 및 캘린더뷰 기능 문제 해결
1. 현재 연도와 월의 데이터들은 잘 기록됨
2. 피커로 연도와 월을 바꿀시 캘린더뷰에 지난 기록들이 반영이 안됨 -> 해결
3. backgroundclearview -> 통계뷰, 캘린더 뷰에서 삭제 후 PickerView로 분리
4. RecordView에서 CalendarVew, RecordElementView로 분리
5. RecordView의 피커 Pickerview로 이동
