+++ Quintet-iOS/View/CalendarPickerView.swift
@@ -1,0 +1,20 @@
+//
+//  CalendarPickerView.swift
+//  Quintet-iOS
+//
+//  Created by 옥재은 on 2023/11/10.
+//
+
+import SwiftUI
+
+struct CalendarPickerView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct CalendarPickerView_Previews: PreviewProvider {
+    static var previews: some View {
+        CalendarPickerView()
+    }
+}
