+++ Quintet-iOS/View/PickerView.swift
@@ -125,6 +125,123 @@ struct MonthPicker: View {
     }
 }
 
+struct CalendarMonthPicker: View {
+    @Binding var selectedMonth: Int
+    @ObservedObject var viewModel: DateViewModel
+
+    var body: some View {
+        Picker("월 선택", selection: $selectedMonth) {
+            ForEach(1...getMaxMonth(), id: \.self) { month in
+                Text("\(month)월")
+            }
+
+        }
+        .pickerStyle(WheelPickerStyle())
+    }
+
+    private func getMaxMonth() -> Int {
+        let currentYear = Calendar.current.component(.year, from: Date())
+        let currentMonth = Calendar.current.component(.month, from: Date())
+        if viewModel.selectedYear == currentYear {
+            if viewModel.selectedMonth > currentMonth {
+                viewModel.selectedMonth = currentMonth
+            }
+            return currentMonth
+        }
+        else {
+            return 12
+        }
+    }
+}
+
+struct CalendarYearPicker: View {
+    @StateObject var viewModel: DateViewModel
+    @Binding var selectedYear: Int
+
+    let currentYear: Int = Calendar.current.component(.year, from: Date())
+
+    var body: some View {
+        Picker("년도 선택", selection: $selectedYear){
+            ForEach(2017...currentYear, id: \.self) { year in
+                Text("\(Utilities.formatNum(year))년")
+            }
+        }
+        .pickerStyle(WheelPickerStyle())
+    }
+}
+
+struct CalendarYearMonthPickerPopup: View {
+    @ObservedObject var viewModel: DateViewModel
+    @Binding var isShowPopup: Bool
+    
+    init(viewModel: DateViewModel, isShowPopup: Binding<Bool>) {
+        self.viewModel = viewModel
+        _isShowPopup = isShowPopup
+    }
+    
+    var body: some View {
+        GeometryReader{ geometry in
+            VStack{
+                HStack(spacing: -7) {
+                    CalendarYearPicker(viewModel: viewModel, selectedYear: $viewModel.selectedYear)
+                    CalendarMonthPicker(selectedMonth: $viewModel.selectedMonth, viewModel: viewModel)
+                }
+            }
+            .padding(.top, 60)
+            HStack(spacing: 10){
+                Button(action: {
+                    isShowPopup = false
+                    viewModel.updateCalendar()
+                }) {
+                    RoundedRectangle(cornerRadius: 10)
+                        .fill(Color("Background"))
+                        .frame(width: 100, height: 40)
+                        .overlay(
+                            Text("취소")
+                                .foregroundColor(.black)
+                                .font(.system(size: 15))
+                        )
+                }
+                Button(action: {
+                    isShowPopup = false
+                    viewModel.selectedYear = viewModel.selectedYear
+                    viewModel.selectedMonth = viewModel.selectedMonth
+                    viewModel.updateCalendar()
+                }) {
+                    RoundedRectangle(cornerRadius: 10)
+                        .fill(Color("DarkQ"))
+                        .frame(width: 100, height: 40)
+                        .overlay(
+                            Text("확인")
+                                .foregroundColor(.white)
+                                .font(.system(size: 15))
+                        )
+                    
+                }
+            }
+            .padding(.top, 300)
+            .padding(.horizontal, 30)
+            
+        }
+        .padding()
+        .frame(width: 300, height: 400)
+        .background(Color.white)
+        .cornerRadius(30)
+        
+    }
+ }
+
+extension Date {
+    func getAllDates() -> [Date] {
+        let calendar = Calendar.current
+        let startDate = calendar.date(from: Calendar.current.dateComponents([.year, .month], from: self))!
+        let range = calendar.range(of: .day, in: .month, for: startDate)!
+        return range.compactMap { day -> Date in
+            return calendar.date(byAdding: .day, value: day - 1, to: startDate)!
+        }
+    }
+}
+
 // MARK: - Utilities
 struct Utilities {
     static func formatNum(_ num: Int) -> String {
@@ -144,4 +261,15 @@ struct Utilities {
         formatter.dateFormat = "MM.dd"
         return formatter.string(from: date)
     }
+}
+
+struct BackgroundClearView: UIViewRepresentable {
+    func makeUIView(context: Context) -> UIView {
+        let view = UIView()
+        DispatchQueue.main.async {
+            view.superview?.superview?.backgroundColor = .clear
+        }
+        return view
+    }
+    func updateUIView(_ uiView: UIView, context: Context) {}
 }
