+++ Quintet-iOS/View/RecordElementView.swift
@@ -1,0 +1,20 @@
+//
+//  RecordElementView.swift
+//  Quintet-iOS
+//
+//  Created by 옥재은 on 2023/10/18.
+//
+
+import SwiftUI
+
+struct RecordElementView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct RecordElementView_Previews: PreviewProvider {
+    static var previews: some View {
+        RecordElementView()
+    }
+}
