+++ Quintet-iOS/View/StatisticsView.swift
@@ -339,18 +339,6 @@ struct DateOptionView: View {
     }
 }
 
-// MARK: - BackgroundClearView
-struct BackgroundClearView: UIViewRepresentable {
-    func makeUIView(context: Context) -> UIView {
-        let view = UIView()
-        DispatchQueue.main.async {
-            view.superview?.superview?.backgroundColor = .clear
-        }
-        return view
-    }
-    func updateUIView(_ uiView: UIView, context: Context) {}
-}
-
 // MARK: - StatisticsCellView
 struct StatisticsCellView: View {
     @Binding var selectedOption: Int
