+++ Quintet-iOS/ViewModel/DateViewModel.swift
@@ -85,5 +85,6 @@ class DateViewModel: ObservableObject {
 
         return days
     }
+    
 }
